requirements.txt
tensorflow
numpy
pillow
opencv-python-headless
flask
streamlit